[compiler] Add `Switch` to `IR` (#13963)
We have no high-level IR analogue to `CodeBuilderLike.switch`. Such a
node is useful for flattening the IR in deeply-nested `If` nodes,
predicated on integer equality.
This partially addresses the stack-overflow error on the
`matrix_muluti_write_nothing` benchmark, which currently has a
stack-overflow error when computing the type of the CDA.