[query] Use apache commons lang3 instead of lang (#13970)
`org.apache.commons.lang` is from the `commons-lang` library, but in
`build.gradle` we explicitly depend on `commons-lang3` which puts
everything under the `lang3` package. We must be picking up
`commons-lang` as some transitive dependency but we no longer get it in
Spark 3.4. Regardless, better to use what we explicitly depend on.