[compiler] Minor Requiredness Performance Enchancements (#13991)
Main change: add `var mark: Int` to `BaseIR`.
On profiling the benchmark `matrix_multi_write_nothing`, I noticed a
significant amount of time was spent
- iterating through zipped arrays in requiredness 
- Adding and removing elements from `HashSet`s.
In fact, half the time spent in requiredness was removing ir nodes from
the `HashSet` set used as the queue! With this change, requiredness runs
like a stabbed rat!

Explanation of `mark`:
This field acts as a flag that analyses can set. For example:
- `HasSharing` can use the field to see if it has visited a node before.
- `Requiredness` uses this field to tell if a node is currently
enqueued.

The `nextFlag` method in `IrMetadata` allows for analyses to get a fresh
value they can set the `mark` field.
This removes the need to traverse the IR after analyses to re-zero every
`mark` field.