[batch] Rename batches tables to job groups (#13810)
Stacked on #13475. This PR renames the following tables to have job
groups in the name instead of batches. Note that this PR needs to
shutdown the batch deployment (offline migration). I'm not 100% sure
this is necessary, but I want to avoid a case where MJC of the database
migration job cannot happen thus deadlocking the system.

```sql
RENAME TABLE batch_attributes TO job_group_attributes,
             batches_cancelled TO job_groups_cancelled,
             batches_inst_coll_staging TO job_groups_inst_coll_staging, 
             batch_inst_coll_cancellable_resources TO job_group_inst_coll_cancellable_resources, 
             aggregated_batch_resources_v2 TO aggregated_job_group_resources_v2,
             aggregated_batch_resources_v3 TO aggregated_job_group_resources_v3, 
             batches_n_jobs_in_complete_states TO job_groups_n_jobs_in_complete_states;
```