[query/vds combiner] A valid vds combiner object can be done
If a finished VariantDatasetCombiner is loaded, it will have no vds
paths or gvcf paths. Before this change, this would cause the loading to
fail, especially in the logic to load extant plans in `new_combiner` for
resuming in progress combines. This will lead to the plan being
regenerated, potentially causing work to be done again. Remove that
sanity check in `__init__` as a nearly identical one exists in
`new_combiner`.

Also, add the string of the exception to the warning about invalid plans
when loading a combiner in `new_combiner`.