[hailctl] make hailctl work on windows (#14090)
CHANGELOG: Fix #14089, which makes `hailctl dataproc connect` work in
Windows Subsystem for Linux.

1. Non 64-bit Windows uses "Program Files" not "Program Files (x86)"

2. Windows Subsystem for Linux looks like GNU/Linux but will not have
chromium on its path.

3. The removed arguments are no longer supported. They produce a warning
message in my version of Chrome and appear to not work in the version of
Chrome that this user was using. Instead, I bind to 0.0.0.0 and access
the Notebook using the machine DNS name. This is how Google recommend
accessing the Spark UI anyway.