[hailtop] Remove namespace options in hailctl auth commands (#14069)
I'm trying to move our codebase away from relying on the notion of a K8s
namespace for routing because this model does not hold in alternative
deployment environments such as Terra. Further, having a `-n` option in
`hailctl auth` commands is awkward because it can only be used for dev
environments yet is visible to users in the help menu. As such, this is
a breaking change only for developers. The functionality is not really
gone though because you can replace any `hailctl auth … -n dgoldste`
with `HAIL_DEFAULT_NAMESPACE=dgoldste hailctl auth …`.