[query] cleave Backend paths into normal parallelize and parallelize returning errors (#14085)
The original goal of this PR was avoiding `Try` when we are not using
the restartability provided by semantic hashing because I strongly
suspect it is related to the loss of stacktraces in exceptions.

Unrelatedly, we realized the semantic hash PR changed the semantics of
Query-on-Spark even when semantic hash is disabled: previously we would
abort RDD writing on the first exception. In Hail 0.2.123 through
0.2.126, the semantics were changed to only crash *after* we already ran
every other partition. Two bad scenarios of which I can think:

1. Suppose the first partition fails due to OOM. We now waste time/money
on the rest of the partitions even though we cannot possibly get a valid
output.

2. Suppose every partition hits a permission error. Users should get
that feedback after paying for O(1) partitions run, not O(N).

I created two Backend paths: the normal `parallelizeAndComputeWithIndex`
with its pre-0.2.123 semantics as well as
`parallelizeAndComputeWithIndexReturnAllErrors` which, as the name says,
returns errors instead of raising them.

While making this change, I think I found two other bugs in the "return
all errors" path, only one of which I addressed in this PR:

1. I'm pretty sure semantic-hash-enabled QoB batch submission is broken
because it uses the logical partition ids as job indices. Suppose there
are 10,000 partitions, but we only need to compute 1, 100, and 1543.
0.2.126 would try to submit a batch of size 3 but whose job indices are
1, 100, and 1543.

2. Likewise, the Query-on-Spark path returns an invalid
`SparkTaskContext.partitionId` which, at best, produces confusing
partition filenames.

I only fixed the former because it was simple to fix. I wasn't exactly
sure what to do about the latter. We should fix that separately because
the changes in this PR need to urgently land in the next release to
avoid unexpected cost when one partition fails.