[doctest] reduce doctest setup time
Before this change, testing just the n_partitions method of Table takes nearly a minute, 52s of
which is spent in "setup". Admittedly, this setup is shared across multiple tests, but this is an
unacceptable burden for iterating on one method.

```
52.30s setup    hail/table.py::hail.table.Table.n_partitions
3.07s call     hail/table.py::hail.table.Table.n_partitions
```

After this change, the setup time significantly reduces. The call gets slower, presumably because
the JVM is not warm. I think the setup time is now dominated by Hail JVM initialization.

```
11.77s call     hail/table.py::hail.table.Table.n_partitions
9.68s setup    hail/table.py::hail.table.Table.n_partitions
```
This reduces the practical runtime of this test by 50%.

This commit adds 72kB to the repository:

```
$ git diff-tree -r -c -M -C --no-commit-id HEAD | awk '{print $4}' | git cat-file --batch-check | awk 'BEGIN {s=0} {s+= $3} END {print s}'
72998
```