[compiler] Code-Generated TableIntervalJoin(..., product = true) (#13617)
Add staged code generation capabilities by lowering to and emitting
`StreamLeftIntervalJoin`.
The `EmitStream` rule for this node uses a code-generated min heap* of
intervals that contain the current key in the left stream, ordered by
right endpoint. 
The row type of the join result is the left row with an array of
values**
from the right table, ordered by their associated interval's right
endpoint.

Note that this implementation is by no means optimised. There are a
number of
opportunities that I'd like to consider in subsequent changes,
including:
- Laziness in restoring the heap property on push/pop from heap.
- Not deep-copying elements of the right stream when no explicit memory
  management per element is required.


\* This min heap is code generated instead of using an off-the-shelf
implementation as:
- we dont yet to have a mapping from `SType` to a java class or
interface to
  parameterise the heap with
- not obvious how to handle region memory management in an existing
solution

\** The value is a row with the key fields omitted.

Additional Changes:
- Make `this` an explicit argument to `CodeBuilder.invoke()`
  - Gives us control which object we're dispatching on.
  - Useful for generating more self-contained classes
  - Previously assumed all methods were defined in the same class.
- Removed `referenceGenomes` from `ExecuteContext`
- Delegate to `Backend`; in practice these come from the backend anyway.
  - These were being populated from the backend object
- Backend is mutable meaning we can add/remove fake genomes for testing