[hailtop] Add subpath to deploy config (#14056)
Currently, the k8s namespace field is used both for routing internal
requests inside kubernetes but also external requests over the internet.
It also has special logic based on whether the namespace indicates a
production or dev environment. For example, if `namespace == 'default'`,
then we route external `batch` requests to `batch.<domain>/path`, but if
`namespace == foo_dev`, we route external `batch` requests to
`internal.<domain>/foo_dev/path`.

This PR decouples the namespace field from routing. Aside from being
overall more straightforward in my opinion, this is necessary for batch
on azure terra where batch is served out of a subpath it does not
control and is unrelated to whatever namespace it might reside in. The
guiding principle for routing is then as follows: If the config has no
subpath, use a subdomain, otherwise put everything under domain +
subpath.

For example:
- `{'domain': 'hail.is', 'subpath': null}` => `batch.hail.is`
- `{'domain': 'internal.hail.is', 'subpath': '/foo_dev'}` =>
`internal.hail.is/foo_dev/batch`

Since the CI pipeline runs on current production instances, there is a
minor need to stay compatible with old deploy configs (or else hack up
the CI build.yaml). It's quite a simple translation though, because if
there is no subpath provided we can infer one based on the
`default_namespace`.