[hailctl] Fix uploading files from parent directories of cwd (#13812)
The bug in #13785 was that we used a `relpath`, but the files that were
to be included was the parent directory of the current working directory
of the script. To solve this, I added a new option `--mounts` that uses
the Docker syntax for `-v src:dest` to specify where the contents of the
source should be mounted into the container.


Fixes #13785