[httpx][Makefile][copier] add --timeout to copier; use in Makefile
We have long had trouble with copying on flaky Internet connections. AFAIK, this is due to our very
aggressive timeouts. This change permits the CLI to override the default timeout in
`hailtop.aiotools.copy`. Setting this back to 600s should work on most flaky Internet
connections. It works on my particularly bad home WiFi.

The upload-docs target was old and broken so I deleted it. I did not change the upload-artifacts
target because that is Google Cloud specific anyway.