[query] better error message when info array field has missing elements (#14105)
Fixes #13346.

Another user was confused by this:
https://github.com/hail-is/hail/issues/14102. Unfortunately, the world
appears to have embraced missing values in VCF array fields even though
the single element case is ambiguous. In #13346, I proposed a scheme by
which we can disambiguate many of the cases, but implementing it ran
into challenges because LoadVCF.scala does not expose whether or not an
INFO field was a literal "." or elided entirely from that line.

Anyway, this error message actually points users to the fix. I also
changed some method names such that every method is ArrayType and never
TypeArray.