[services] Fix domain configuration in dev namespaces (#14164)
As of #14056, there is ambiguity when referring to the configuration
value `domain`.

- In the `global-config`, this is the root domain of the entire hail
system. This is the same across all namespaces.
- In the `deploy-config` of a namespace N, this refers to the root
domain served by applications of that namespace. In production
(namespace `default`), this is `hail.is`, the same as the root domain of
the entire system. In other namespaces, this is `internal.hail.is`.

Setting the `HAIL_DOMAIN` environment variable in the k8s deployments
from the `global-config` overrides what should be `internal.hail.is` to
`hail.is`, breaking any form of redirection. There's really no need to
set this environment variable at all, as its value can be derived from
the `deploy-config`. This PR removes that environment variable. I tested
this in my development namespace.