[hailctl] Dev config set should only change one property (#14169)
#14056 added a new optional field to the deploy config, `base_path`
which is intended to phase out `default_namespace`. But for backwards
compatibility reasons we cannot yet remove `default_namespace`. This
should all work fine without breaking any workflows like switching back
and forth between namespaces so long as `base_path` is not explicitly
set in a developer's deploy config.

But `hailctl dev config set <property> <value>` does not just set a
single property, it loads the deploy config, sets the property, and then
writes the whole deploy config back. If the deploy config does anything
with default values, which it now does with `base_path`, this round trip
does not work.

Another simpler example is that currently in main, the following will
make two changes to a deploy config not one:

```
# deploy config of {'location': 'external', 'domain': 'hail.is', 'default_namespace': 'default'}
HAIL_DOMAIN=foo hailctl dev config set location gce

# deploy config will now read {'location': 'gce', 'domain': 'foo', 'default_namespace': 'default'}
```

This PR should change `hailctl dev config set` so that the only changes
that are made to the deploy config are the single property/value change
described in the command.