[monitoring] Use aiohttp.AppKey to add static typing to Application storage (#14173)
Forgive my throwing the dev proxy change in here too, but this gets us
static typechecking on stuff stored in `app` across the codebase except
for `batch`. That's a beast that will be a lot of changes to port to
`AppKey`s and I don't want to mess with the current in flight PRs.