[batch] Add metadata server to batch jobs in GCP (#14019)
Implements a basic GCP metadata server for user jobs as described in
https://github.com/hail-is/hail-rfcs/pull/12. It implements only so much
as is needed for `hail` and `gcloud` to get access tokens for hail GSAs
so they can then make API calls to GCS or Hail Batch. With this in place
user jobs should no longer require GSA key files, but removing them is
future work and requires a well-communicated deprecation and removal
process.