[query] Use valid globals reference in MWZJ and TABK (#14246)
CHANGELOG: Fix a bug, introduced in 0.2.114, in which
`Table.multi_way_zip_join` and `Table.aggregate_by_key` could throw
"NoSuchElementException: Ref with name `__iruid_...`" when one or more
of the tables had a number of partitions substantially different from
the desired number of output partitions.

Fixes https://github.com/hail-is/hail/issues/14245.

In both MultiWayZipJoin and TableAggregateByKey, we repartition the
child but neglect to use the new globals `Ref` from the repartitioned
child. As long as `repartitionNoShuffle` does not create a new
TableStage with new globals, this is fine, but that is not, in general,
true. It seems that recently, in lowered backends, when the repartition
cost is deemed "high" we generate a fresh TableStage with a fresh
globals ref.