[query] support importing empty JSON objects (#14202)
@patrick-schultz I'm not sure if this makes sense or not, but I observed
it while working on something else. It seems weird but acceptable to
import an empty dictionary as any struct. Does this seem reasonable to
you? How have we avoided this bug for so long?

I'm not familiar enough with this code to know how to simply reproduce
the bug and add a corresponding test. Thoughts?