Fix an error in the MatrixTable tutorial (#14239)
### Description

In this pull request, I fix an error in the MatrixTable tutorial. The
tutorial shows some genotype data and erroneously states that all the
genotypes that are shown are homozygous reference (0/0). In fact, there
are also some heterozygous (0/1) and homozygous alternate (1/1)
genotypes in the displayed data. In this pull request, I remove the
erroneous statement.

### Testing

I ran the notebook to confirm that the notebook displays a mix of
genotypes, not just homozygous reference. You can view the erroneous
version of the tutorial
[here](https://hail.is/docs/0.2/tutorials/07-matrixtable.html#MatrixTable-operations).