[gear] Change deadlock logs from warning to info (#14251)
As discussed in #14240, we emit warnings on database deadlocks, which
there are enough of to trigger noisy alerts. Since there's nothing to be
done operationally (and there's no current work underway to get rid of
them), these alerts only contribute to alert fatigue and hide potential
problems in the system that could be addressed. This demotes a deadlock
to the `info` level so we can still see how often they occur but are not
alerted by them. In the future when we resolve the current deadlock we
can re-escalate this error so that we can catch new deadlocks that are
introduced.