Update Table method examples (#14242)
### Description

In this pull request, I update the docstrings for the `sample`, `head`,
and `semi_join` methods of the `Table` class with clearer examples that
demonstrate the functionality of each method.

I also update the unit test for the `sample` method so that it is
deterministic and stronger.

### Testing

I ran `make -C hail doctest-query hail/table.py` to confirm that the
docstring tests are passing for the methods that I updated.