[infra] Make CI an admin of the artifact registry (#14266)
If CI jobs are to apply cleanup policies to the Artifact Registry, CI
needs to have the permissions on the AR to do so. Prior to this change,
CI only had viewer on the AR and used a different service account named
`gcr-push` to push images to the registry. I don't see the point in
having this second service account and think that CI should have these
permissions (because it basically does already if it can submit jobs
with the `gcr-push` service account). So the main point of this PR is to
give CI admin on the AR, and then we can follow up by deleting the
`gcr-push` service account.

Separately:

Not sure how this ever worked, but I needed to update the google
terraform provider to get `cleanup_policy_dry_run = false`. That upgrade
now made the DNS zone description mandatory so I added one. I thin the
DNS zone is no longer used, but removing it is for another time. There
was also for some reason a duplicate resource for CI's viewer role on
the AR repo, so I changed one to admin and deleted the other.

cc @ehigham

---------

Co-authored-by: Dan King <dking@broadinstitute.org>