[release] remove unused old stuff (#14273)
cloudtools is an ancient name for hailctl dataproc. This is not the docs
SHA. The docs are copied into a Docker image by the `website_image`
build step. The only special file is
gs://hail-common/builds/0.1/docs/hail-0.1-docs-5a6778710097.tar.gz which
was the last 0.1 docs. It is the one that we host on the website for
posterity.