update test-gcp.sh to be closer to modern use (#14272)
AFAIK, we do not use this file, but it is a nice historical note of how
to run the Hail tests on a Dataproc cluster, if we ever wanted to do
that. These days we only run a limited set of tests via test-dataproc.sh

I am also open to deleting this file since it certainly does not work
now.