[fs] Refresh blob client on every range read (#14278)
In anticipation of terra where we will need to refresh SAS tokens with
WSM, this change grabs a blob client on every range read instead of
stashing one from the first read. This gives the opportunity for
`get_blob_client` to refresh expired credentials and create a new blob
client in circumstances where we are streaming through large data.