[annotationdb][datasets] regional buckets (#14286)
CHANGELOG: In GCP, the Hail Annotation DB and Datasets API have moved
from multi-regional US and EU buckets to regional US-CENTRAL1 and
EUROPE-WEST1 buckets. These buckets are requester pays which means
unless your cluster is in the US-CENTRAL1 or EUROPE-WEST1 region, you
will pay a per-gigabyte rate to read from the Annotation DB or Datasets
API. We must make this change because [reading from a multi-regional
bucket into a regional VM is no longer
free](https://cloud.google.com/storage/pricing-announce#network).
Unfortunately, cost constraints require us to choose only one region per
continent and we have chosen US-CENTRAL1 and EUROPE-WEST1.

This is a disruptive change.

Due to the introduction of egress costs from multi-regional buckets, we
decided to move all the datasets in the Annotation DB and Datasets API
into a regional bucket. This both reduces costs for us (regional is 77%
of multi-regional) and reduces costs for users (data leaving a US
multi-regional bucket and entering any north american region is 0.02,
data leaving a US-CENTRAL1 bucket entering a US-CENTRAL1 VM is free).

I already copied the US data into the new regional buckets. The european
data is in progress. We should delete the multi-regional buckets as soon
as possible given the increased costs, but I will not remove them until
this PR is merged and then released.

I will announce this change on @hailgenetics twitter and
#hail-announcements on ATGU slack.

Also notice that `hail-uk-vep` was already regional but the name did not
agree with the new naming scheme so I moved it too.

| old | new |
| --- | --- |
| hail-datasets-eu | hail-datasets-europe-west1 |
| hail-datasets-us | hail-datasets-us-central1	|
| hail-eu-vep      | hail-europe-west1-vep	|
| hail-us-vep      | hail-us-central1-vep       |
| hail-uk-vep      | hail-europe-west2-vep      |


cc: @chrisvittal , @patrick-schultz for visibility into this potentially
disruptive Query change.