[ci] Use the public networks where possible (#14294)
`network=private` is an escape hatch for CI so certain jobs can talk to
internal endpoints on our network that we do not permit user jobs to
reach. In `main`, all CI jobs are hard-coded to use `private` in
`build.py`, but few jobs in the CI pipeline actually require these
heightened privileges. The steps in the CI pipeline are of the following
types:

- `BuildImageStep`: These do not need to use the private network and
have now all been made to use the public network namespaces
- `CreateDatabaseStep`: These *do* need to use `network='private'`
because all our DBs only have private IPs on our internal network
- `RunImageStep`: Those steps that contact the DB need the private
network. This PR makes the network configurable for these steps but
default to public, so steps that need DB access explicitly do `network:
private`
- `DeployStep`: These do not need to use the private network because
they use the public K8s API server endpoint. Whether they should is
perhaps a different question. I'm open to keeping these on the private
networks and creating an issue to use the internal API server endpoint
instead. We definitely have a static internal IP in GKE but I don't
believe we have one for AKS and that would involve some research.
- `CreateNamespaceStep`: I don't believe that this needs the private
network because it is functionally the same as a `DeployStep` in that it
just talks to K8s, but I am unable to test this step in `test_ci` so I
am reluctant to make a change that could brick CI. I instead made it
configurable and default to its current value ('private'). We could then
make a follow-up PR that tries turning it public.